move resourceinput to separate file
