+++ src/rsrc_mgr/resourceinput.h
@@ -1,0 +1,169 @@
+#ifndef _RESOURCEINPUT_H
+#define _RESOURCEINPUT_H
+/*
+Copyright (c) 2015, Los Alamos National Security, LLC
+All rights reserved.
+
+Copyright 2015.  Los Alamos National Security, LLC. This software was
+produced under U.S. Government contract DE-AC52-06NA25396 for Los
+Alamos National Laboratory (LANL), which is operated by Los Alamos
+National Security, LLC for the U.S. Department of Energy. The
+U.S. Government has rights to use, reproduce, and distribute this
+software.  NEITHER THE GOVERNMENT NOR LOS ALAMOS NATIONAL SECURITY,
+LLC MAKES ANY WARRANTY, EXPRESS OR IMPLIED, OR ASSUMES ANY LIABILITY
+FOR THE USE OF THIS SOFTWARE.  If software is modified to produce
+derivative works, such modified software should be clearly marked, so
+as not to confuse it with the version available from LANL.
+
+Additionally, redistribution and use in source and binary forms, with
+or without modification, are permitted provided that the following
+conditions are met: 1. Redistributions of source code must retain the
+above copyright notice, this list of conditions and the following
+disclaimer.
+
+2. Redistributions in binary form must reproduce the above copyright
+notice, this list of conditions and the following disclaimer in the
+documentation and/or other materials provided with the distribution.
+3. Neither the name of Los Alamos National Security, LLC, Los Alamos
+National Laboratory, LANL, the U.S. Government, nor the names of its
+contributors may be used to endorse or promote products derived from
+this software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY LOS ALAMOS NATIONAL SECURITY, LLC AND
+CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING,
+BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
+FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL LOS
+ALAMOS NATIONAL SECURITY, LLC OR CONTRIBUTORS BE LIABLE FOR ANY
+DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE
+GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR
+OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
+ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+-----
+NOTE:
+-----
+MarFS is released under the BSD license.
+
+MarFS was reviewed and released by LANL under Los Alamos Computer Code
+identifier: LA-CC-15-039.
+
+MarFS uses libaws4c for Amazon S3 object communication. The original version
+is at https://aws.amazon.com/code/Amazon-S3/2601 and under the LGPL license.
+LANL added functionality to the original work. The original work plus
+LANL contributions is found at https://github.com/jti-lanl/aws4c.
+
+GNU licenses can be found at http://www.gnu.org/licenses/.
+*/
+
+#include <pthread.h>
+#include <stdint.h>
+
+#include "config/config.h"
+#include "mdal/mdal.h"
+#include "rsrc_mgr/consts.h"
+#include "rsrc_mgr/resourcelog.h"
+
+typedef struct {
+   // synchronization and access control
+   pthread_mutex_t  lock;     // no simultaneous access
+   pthread_cond_t   complete; // signaled when all work is handed out
+   pthread_cond_t   updated;  // signaled when new work is added, or input state is otherwise changed
+   char             prepterm; // flag value set when clients should prepare for termination
+   size_t           clientcount; // count of clients still running
+   // previous log info
+   RESOURCELOG      rlog;
+   // state info
+   marfs_ns*        ns;
+   MDAL_CTXT        ctxt;
+   // reference info
+   ssize_t          refindex;
+   ssize_t          refmax;
+}*RESOURCEINPUT;
+
+/**
+ * Initialize a given resourceinput
+ * @param RESOURCEINPUT* resourceinput : Resourceinput to be initialized
+ * @param marfs_ns* ns : MarFS NS associated with the new resourceinput
+ * @param MDAL_CTXT ctxt : MDAL_CTXT associated with the previous NS
+ *                         NOTE -- caller should never modify this again
+ * @return int : Zero on success, or -1 on failure
+ */
+int resourceinput_init( RESOURCEINPUT* resourceinput, marfs_position* pos, size_t clientcount );
+
+/**
+ * Update the given RESOURCEINPUT structure to use the given logpath as a new input source
+ * @param RESOURCEINPUT* resourceinput : Resourceinput to update
+ * @param const char* logpath : Path of the new (RECORD) resourcelog
+ * @return int : Zero on success, or -1 on failure
+ */
+int resourceinput_setlogpath( RESOURCEINPUT* resourceinput, const char* logpath );
+
+/**
+ * Set the active reference dir range of the given resourceinput
+ * NOTE -- this will fail if the range has not been fully traversed
+ * @param RESOURCEINPUT* resourceinput : Resourceinput to have its range set
+ * @param ssize_t start : Starting index of the reference dir range
+ * @param ssize_t end : Ending index of the reference dir range ( non-inclusive )
+ * @return int : Zero on success, or -1 on failure
+ */
+int resourceinput_setrange( RESOURCEINPUT* resourceinput, size_t start, size_t end );
+
+/**
+ * Get the next ref index to be processed from the given resourceinput
+ * @param RESOURCEINPUT* resourceinput : Resourceinput to get the next ref index from
+ * @param opinfo** nextop : Reference to be populated with a new op from an input logfile
+ * @param MDAL_SCANNER* scanner : Reference to be populated with a new reference scanner
+ * @param char** rdirpath : Reference to be populated with the path of a newly opened reference dir
+ * @return int : Zero, if no inputs are currently available;
+ *               One, if an input was produced;
+ *               Ten, if the caller should prepare for termination ( resourceinput is preparing to be closed )
+ */
+int resourceinput_getnext( RESOURCEINPUT* resourceinput, opinfo** nextop, MDAL_SCANNER* scanner, char** rdirpath );
+
+/**
+ * Destroy all available inputs and signal threads to prepare or for imminent termination
+ * NOTE -- this is useful for aborting, if a thread has hit a fatal error
+ * @param RESOURCEINPUT* resourceinput : Resourceinput to purge
+ * @param int : Zero on success, or -1 on failure
+ */
+int resourceinput_purge( RESOURCEINPUT* resourceinput );
+
+/**
+ * Wait for the given resourceinput to have available inputs, or for immenent termination
+ * @param RESOURCEINPUT* resourceinput : Resourceinput to wait on
+ * @param int : Zero on success, or -1 on failure
+ */
+int resourceinput_waitforupdate( RESOURCEINPUT* resourceinput );
+
+/**
+ * Wait for the given resourceinput to be terminated ( synchronizing like this ensures ALL work gets enqueued )
+ * @param RESOURCEINPUT* resourceinput : Resourceinput to wait on
+ * @param int : Zero on success, or -1 on failure
+ */
+int resourceinput_waitforterm( RESOURCEINPUT* resourceinput );
+
+/**
+ * Wait for all inputs in the given resourceinput to be consumed
+ * @param RESOURCEINPUT* resourceinput : Resourceinput to wait on
+ * @param int : Zero on success, or -1 on failure
+ */
+int resourceinput_waitforcomp( RESOURCEINPUT* resourceinput );
+
+/**
+ * Terminate the given resourceinput
+ * @param RESOURCEINPUT* resourceinput : Resourceinput to be terminated
+ * @return int : Zero on success, or -1 on failure
+ */
+int resourceinput_term( RESOURCEINPUT* resourceinput );
+
+/**
+ * Destroy the given resourceinput
+ * @param RESOURCEINPUT* resourceinput : Resourceinput to be destroyed
+ * @return int : Zero on success, or -1 on failure
+ */
+int resourceinput_destroy( RESOURCEINPUT* resourceinput );
+
+#endif
